+++ CHANGELOG.md
@@ -1,9 +1,42 @@
 # Change Log
 
-All notable changes to the "TinyDiagram" extension will be documented in this file.
+All notable changes to the "Code Prism" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### 1.2.0 - November 3, 2024
 
-## 0.0.1 - September 29, 2020
+1. implement to cahnge issue's position
+2. implement to change note's category
+3. implement to make markdown file related the note
+
+### 1.1.0 - October 31, 2024
+
+> 🚨 This version is not compatible with the previous versions  
+>  (if you need your data, please rename the `context` to `content` in all prism files in your prism folder(it is normally `<your project>.prism`) T.T)
+
+- improve function about link in comment controller's comment
+
+### 1.0.8 - October 31, 2024
+
+- improve the link function
+
+### 1.0.7 - October 30, 2024
+
+- fix the bug that don't show the linked file's context in hover tooltip
+
+### 1.0.6 - October 30, 2024
+
+- changed the extension's icon
+
+### 1.0.5 - October 30, 2024
+
+1. fix the bug that is malfunction of cancel button when input issue in comment controller.
+2. implement to show the link information when comment is shown in comment controller.
+3. and so on
+
+### 1.0.4 - October 29, 2024
+
+- bug fixed : fixed the bug that couldn't search the documents relatived source code
+
+### 1.0.0 - October 29, 2024
 
 - Initial release